Move initial branchDiff to prepare
